+++ tests_python/tests_alpha/test_mockup.py
@@ -495,6 +495,17 @@ def _test_create_mockup_init_show_roundtrip(
     ba_sent = _try_json_loads(_BA_FLAG, ba_str)
     pc_sent = _try_json_loads(_PC_FLAG, pc_str)
 
+    # Test that the initial mockup call honored the values it received. If
+    # it didn't, all calls would return the default values all along, and
+    # everything would seem fine; but it wouldn't be. This was witnessed in
+    # https://gitlab.com/tezos/tezos/-/issues/938
+    if bootstrap_json:
+        ba_input = json.loads(bootstrap_json)
+        assert ba_sent == ba_input
+    if protocol_constants_json:
+        pc_input = json.loads(protocol_constants_json)
+        assert pc_sent == pc_input
+
     # 3/ Pass obtained json to a new mockup instance, to check json
     # is valid w.r.t. ocaml encoding
 
@@ -559,8 +570,12 @@ def _test_create_mockup_init_show_roundtrip(
         None,
         json.dumps(
             {
+                'hard_gas_limit_per_operation': "1040001",
+                'hard_storage_limit_per_operation': "60001",
+                'cost_per_byte': "251",
+                'hard_gas_limit_per_block': "10400001",
                 'chain_id': "NetXaFDF7xZQCpR",
-                'initial_timestamp': "2020-07-21T17:11:10+02:00",
+                'initial_timestamp': "2020-07-21T17:11:10Z",
             }
         ),
     ],